+++ app/Services/ProjectService.php
@@ -18,12 +18,15 @@ class ProjectService
      * @var \App\Models\Project
      */
     private $model;
+    private $modelTaskLog;
 
     function __construct(
-        Project $project
+        Project $project,
+        TaskLog $taskLog
     )
     {
         $this->model = $project;
+        $this->modelTaskLog = $taskLog;
     }
 
     function all() {
@@ -33,13 +36,20 @@ class ProjectService
 
     function show($id) {
         $project = $this->model->where('id', $id)->first();
-        $logs = $this->getLogsByProject($id);
-        $project['logs'] = $logs;
+        if(!is_null($project)) {
+            $logs = $this->getLogsByProject($id);
+            $project['logs'] = $logs;
+        }
 
         return $project;
     }
 
     function store($request, $userId) {
+
+        if(!isset($request->name) || !isset($request->description)) {
+            return false;
+        }
+
         $project = new Project();
         $project->name = $request->name;
         $project->description = $request->description;
@@ -51,6 +61,11 @@ class ProjectService
     }
 
     function update($request, $id) {
+
+        if(!isset($request->name) || !isset($request->description)) {
+            return false;
+        }
+
         $project = $this->model->findOrFail($id);
         $project->name = $request->name;
         $project->description = $request->description;
@@ -61,22 +76,31 @@ class ProjectService
     }
 
     function destroy($id) {
+        if(!is_integer($id)) {
+            return false;
+        }
         $project = $this->model->destroy($id);
 
         return $project;
     }
 
     function getMyProjectsByUserID($id) {
+        if(!is_integer($id)) {
+            return false;
+        }
         return $this->getProjectsByUserID($id);
     }
 
     function getProjectsByUserID($id) {
+        if(!is_integer($id)) {
+            return false;
+        }
         return $this->model->with(['Tasks', 'Tasks.Type'])->where('user_id', $id)->get();
     }
 
     function getLogsByProject($id) {
         $project_id = $id;
-        $logs = TaskLog::with(['Task', 'Task.Project', 'Task.Project.User', 'TaskType'])->whereHas('Task', function ($query) use ($project_id) {
+        $logs = $this->modelTaskLog->with(['Task', 'Task.Project', 'Task.Project.User', 'TaskType'])->whereHas('Task', function ($query) use ($project_id) {
             $query->where('project_id', $project_id);
         })->orderBy('created_at', 'desc')->get();
